i965: Expand INTEL_DEBUG to uint64_t.

+++ src/mesa/drivers/dri/i965/intel_debug.c
@@ -33,7 +33,7 @@
 #include "intel_debug.h"
 #include "utils.h"
 
-int INTEL_DEBUG = 0;
+uint64_t INTEL_DEBUG = 0;
 
 static const struct dri_debug_control debug_control[] = {
    { "tex",   DEBUG_TEXTURE},
